cli/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
click==8.1.3
    # via typer
pyyaml==6.0
    # via -r requirements.in
typer==0.4.1
    # via -r requirements.in
